Create UserAvatarConnected component
This includes moving UserAvatarHeader and ChangeAvatar components.
The ChangeAvatar component was renamed to UserAvatarUpdate
