+++ packages/apollos-ui-connected/package.json
@@ -44,7 +44,9 @@
     "react": "*",
     "react-native": "*",
     "react-native-inappbrowser-reborn": "*",
-    "react-navigation": "*"
+    "react-navigation": "*",
+    "react-native-image-picker": "*",
+    "react-native-permissions": "*"
   },
   "dependencies": {
     "graphql-tag": "2.9.2",
@@ -59,6 +61,8 @@
     "react": "16.8.6",
     "react-native": "0.60.5",
     "react-native-inappbrowser-reborn": "3.3.4",
+    "react-native-image-picker": "0.28.0",
+    "react-native-permissions": "2.0.5",
     "react-navigation": "^3.11.0",
     "react-test-renderer": "16.6.3",
     "waait": "1.0.2"
