+++ packages/apollos-ui-connected/src/UserAvatarConnected/UserAvatarHeader/getUserProfile.js
