+++ packages/apollos-ui-connected/src/UserAvatarConnected/UserAvatarHeader/index.js
