+++ packages/apollos-ui-connected/src/UserAvatarConnected/UserAvatarUpdate/index.js
@@ -1,0 +1,3 @@
+import UserAvatarUpdate from './UserAvatarUpdate';
+
+export default UserAvatarUpdate;
