+++ packages/apollos-ui-connected/src/UserAvatarConnected/getUserPhoto.js
