+++ packages/apollos-ui-connected/src/UserAvatarConnected/index.js
@@ -1,0 +1,6 @@
+import UserAvatarConnected from './UserAvatarConnected';
+
+export UserAvatarUpdate from './UserAvatarUpdate';
+export UserAvatarHeader from './UserAvatarHeader';
+
+export default UserAvatarConnected;
